Timeout monitoring system
Implements a subsystem to monitor the number of times a user is timed
out, and the total duration of said timeouts.

**Limitations:**
Any timeouts which occur whilst the bot is offline will not be tracked,
and timeouts which overlap with an offline period _may_ introduce
inaccuracy in the statistics (the entire duration of the timeout is
presumed to take place when the timeout is issued, and the difference is
accounted for if a timeout is prematurely terminated. If, however, the
early end is not received by the bot, then the timeout duration will not
be reduced appropriately.)

This also introduces the `GUILD_MEMBERS` privileged intent as a
requirement.