Apply tracing to all lock uses.
An issue was observed where, if the Memes & Nickname Announcements channels were the same, the nickname announcement would consistently trigger a deadlock.

The cause of this is not very clear, beyond appearing that the waiting write handle from the memes subsystem's New Message handler prevents additional read handles required by the nickname lottery error handling, thereby preventing execution.

This addresses the issue in this case by sharing the existing read handle to ensure that the function can finish (and thereby drop the read handle!).