Rewrite core meme voting logic
Rewritten the meme voting logic thread, so that we no longer have to maintain a local copy of all MessageIDs.

The new logic instead simply preserves an 'initial message ID', and whenever the list of messages is needed this is requested on-demand from Discord.

The intent is to improve network resilience, as we can now better be certain that we haven't missed anything whilst offline.
The new structure also better supports retrying quickly when a network request fails (such as due to no network).

Also drive-by changed all `&ctx.http()` to simply `&ctx` to use the cache.

**SemVer:** Breaking change to config.