Remove duplicate work
In praxis, we shouldn't need to re-do the earlier checks anyway, since the memes processing routine checks for validity too.

There shouldn't be any *normal* reason for the settings to change during the bot's slumber.

(And moving the sleep seems to cause Tokio to use 100% of a CPU core on AArch64 for some reason?)